backend/requirements.txt
flask
flask-cors
google-generativeai
fpdf
python-dotenv

pytest
